# Front/SalesDatePredictionApp/src/app/services/api/models/order-detail-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface OrderDetailDto {
  discount?: number;
  productId?: number;
  qty?: number;
  unitPrice?: number;
}