# Front/SalesDatePredictionApp/src/app/services/api/models/order-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { OrderDetailDto } from '../models/order-detail-dto';
export interface OrderDto {
  empId?: number;
  freight?: number;
  orderDate?: string;
  orderDetails?: Array<OrderDetailDto> | null;
  requiredDate?: string;
  shipAddress?: string | null;
  shipCity?: string | null;
  shipCountry?: string | null;
  shipName?: string | null;
  shippedDate?: string | null;
  shipperId?: number;
}